👏 Update new compatible processor, automatically saved processor pretrained to dump folder.
