+++ .gitignore
@@ -32,3 +32,5 @@ ljspeech
 /datasets
 /examples/tacotron2/exp/
 /temp/
+kss
+
