+++ tensorflow_tts/bin/preprocess.py
@@ -32,6 +32,10 @@ from tqdm import tqdm
 
 from tensorflow_tts.processor import LJSpeechProcessor
 from tensorflow_tts.processor import KSSProcessor
+
+from tensorflow_tts.processor.ljspeech import LJSPEECH_SYMBOLS
+from tensorflow_tts.processor.kss import KSS_SYMBOLS
+
 from tensorflow_tts.utils import remove_outlier
 
 os.environ["CUDA_VISIBLE_DEVICES"] = ""
@@ -122,9 +126,9 @@ def gen_audio_features(item, config):
     # check audio properties
     assert len(audio.shape) == 1, f"{utt_id} seems to be multi-channel signal."
     assert np.abs(audio).max() <= 1.0, f"{utt_id} is different from 16 bit PCM."
-    assert (
-        rate == config["sampling_rate"]
-    ), f"{utt_id} sampling rate is not {config['sampling_rate']}."
+    # assert (
+    #     rate == config["sampling_rate"]
+    # ), f"{utt_id} sampling rate is not {config['sampling_rate']}."
 
     # trim silence
     if config["trim_silence"]:
@@ -256,7 +260,12 @@ def preprocess():
 
     dataset_processor = {
         "ljspeech": LJSpeechProcessor,
-        "kss": KSSProcessor
+        "kss": KSSProcessor,
+    }
+
+    dataset_symbol = {
+        "ljspeech": LJSPEECH_SYMBOLS,
+        "kss": KSS_SYMBOLS,
     }
 
     dataset_cleaner = {
@@ -266,7 +275,7 @@ def preprocess():
 
     logging.info(f"Selected '{config['dataset']}' processor.")
     processor = dataset_processor[config["dataset"]](
-        config["rootdir"], cleaner_names=dataset_cleaner[config["dataset"]]
+        config["rootdir"], symbols=dataset_symbol[config["dataset"]], cleaner_names=dataset_cleaner[config["dataset"]]
     )
 
     # check output directories
@@ -276,6 +285,9 @@ def preprocess():
     ]
     build_dir("train")
     build_dir("valid")
+
+    # save pretrained-processor to feature dir
+    processor._save_mapper(os.path.join(config["outdir"], f"{config['dataset']}_mapper.json"))
 
     # build train test split
     train_split, valid_split = train_test_split(
