+++ tensorflow_tts/processor/base_processor.py
@@ -1,181 +1,119 @@
-# -*- coding: utf-8 -*-
-# Copyright 2020 The TensorFlowTTS Team
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""Abstract class for processor"""
-
-
+from dataclasses import dataclass, field
 import abc
+import os
+import json
+from typing import Union, List, Dict
 
 
-class BaseProcessor(metaclass=abc.ABCMeta):
-    """Base Processor class for all Processor."""
+class DataProcessorError(Exception):
+    pass
 
 
-    SPECIAL_TOKENS_ATTRIBUTES = [
-        "unk",
-        "pad",
-        "eos",
-        "bos",
-    ]
+@dataclass
+class BaseProcessor(abc.ABC):
+    data_dir: str
+    symbols: List[str] = field(default_factory=list)
+    speakers_map: Dict[str, int] = field(default_factory=dict)
+    train_f_name: str = "train.txt"
+    delimiter: str = "|"
+    positions = {
+        "file": 0,
+        "text": 1,
+        "speaker_name": 2,
+    }  # positions of file,text,speaker_name after split line
+    f_extension: str = ".wav"
+    #extra_tokens = {"unk": "[UNK]", "pad": "[PAD]", "eos": "[EOS]", "bos": "[BOS]"}
+    save_mapper: bool = False
+    load_mapper: bool = False
+    # extras
+    items: List[List[str]] = field(default_factory=list)  # text, wav_path, speaker_name
+    symbol_to_id: Dict[str, int] = field(default_factory=dict)
+    id_to_symbol: Dict[int, str] = field(default_factory=dict)
 
-    def __init__(
-        self,
-        items,
-        symbols,
-        unk="[UNK]",
-        pad="[PAD]",
-        eos="[EOS]",
-        bos="[BOS]",
-        **kwargs
-    ):
-        super().__init__(**kwargs)
-        self._unk = unk
-        self._pad = pad
-        self._eos = eos
-        self._bos = bos
+    def __post_init__(self):
 
-        # Update SPECIAL_TOKENS.
-        for key, value in kwargs.items():
-            if key in self.SPECIAL_TOKENS_ATTRIBUTES:
-                if isinstance(value, str):
-                    setattr(self, key, value)
-                else:
-                    raise TypeError(
-                        "special token {} has to be str but got: {}".format(key, type(value))
-                    )
+        if self.load_mapper:
+            self._load_mapper()
+            return
 
-        self._items = None
-        self._symbols = None
+        if self.symbols.__len__() < 1:
+            raise DataProcessorError("Symbols list is empty but mapper isn't loaded")
 
-        self.items = items
-        self.symbols = symbols
+        self.create_items()
+        self.create_speaker_map()
+        self.reverse_speaker = {v: k for k, v in self.speakers_map.items()}
+        self.create_symbols()
+        if self.save_mapper:
+            self._save_mapper()
 
-        # update symbols with special tokens
-        # to keep backward compatibility, pad always a first symbols
-        # and eos always after self_symbols.
-        self.symbols = [self.pad] + self._symbols + [self.eos] + [self.bos] + [self.unk]
-        
-        # symbols mapping
-        self._symbol_to_id = {s: i for i, s in enumerate(self.symbols)}
-        self._id_to_symbol = {i: s for i, s in enumerate(self.symbols)}
+    def __getattr__(self, name: str) -> Union[str, int]:
+        if "_id" in name:  # map symbol to id
+            return self.symbol_to_id[name.replace("_id", "")]
+        return self.symbol_to_id[name]  # map symbol to value
 
-    def add_symbols(self, new_symbols):
-        """Add new symbols to the end of self._symbols."""
-        if isinstance(new_symbols, str):
-            self._symbols = self._symbols + [new_symbols]
-        elif isinstance(new_symbols, list):
-            for s in new_symbols:
-                if isinstance(s, str):
-                    self._symbols = self._symbols + [s]
-                else:
-                    raise ValueError("All elements of new_symbols must be a str.")
+    def create_speaker_map(self):
+        """
+        Create speaker map for dataset
+        """
+        sp_id = 0
+        for i in self.items:
+            speaker_name = i[-1]
+            if speaker_name not in self.speakers_map:
+                self.speakers_map[speaker_name] = sp_id
+                sp_id += 1
+
+    def get_speaker_id(self, name: str) -> int:
+        return self.speakers_map[name]
+
+    def get_speaker_name(self, speaker_id: int) -> str:
+        return self.speakers_map[speaker_id]
+
+    def create_symbols(self):
+        self.symbol_to_id = {s: i for i, s in enumerate(self.symbols)}
+        self.id_to_symbol = {i: s for i, s in enumerate(self.symbols)}
+
+    def create_items(self):
+        """
+        Method used to create items from training file
+        items struct => text, wav_file_path, speaker_name
+        """
+        with open(
+            os.path.join(self.data_dir, self.train_f_name), mode="r", encoding="utf-8"
+        ) as f:
+            for line in f:
+                parts = line.strip().split(self.delimiter)
+                wav_path = os.path.join(self.data_dir, parts[self.positions["file"]])
+                wav_path = (
+                    wav_path + self.f_extension
+                    if wav_path[-len(self.f_extension) :] != self.f_extension
+                    else wav_path
+                )
+                text = parts[self.positions["text"]]
+                speaker_name = parts[self.positions["speaker_name"]]
+                self.items.append([text, wav_path, speaker_name])
+
+    def add_symbol(self, symbol: Union[str, list]):
+        if isinstance(symbol, str):
+            if symbol in self.symbol_to_id:
+                return
+            self.symbols.append(symbol)
+            symbol_id = len(self.symbol_to_id)
+            self.symbol_to_id[symbol] = symbol_id
+            self.id_to_symbol[symbol_id] = symbol
+
+        elif isinstance(symbol, list):
+            for i in symbol:
+                self.add_symbol(i)
         else:
             raise ValueError("A new_symbols must be a string or list of string.")
 
-    @property
-    def unk(self):
-        """Unkow character/phoneme."""
-        return str(self._unk)
-
-    @property
-    def pad(self):
-        """Padding."""
-        return str(self._pad)
-
-    @property
-    def eos(self):
-        """end of sentence."""
-        return str(self._eos)
-
-    @property
-    def bos(self):
-        """begin of sentence."""
-        return str(self._bos)
-
-    @property
-    def items(self):
-        return self._items
-
-    @items.setter
-    def items(self, value):
-        self._items = value
-
-    @property
-    def symbols(self):
-        return self._symbols
-    
-    @symbols.setter
-    def symbols(self, value):
-        self._symbols = value
-
-    @property
-    def symbol_to_id(self):
-        return self._symbol_to_id
-
-    @property
-    def id_to_symbol(self):
-        return self._id_to_symbol
-
-    @unk.setter
-    def unk(self, value):
-        self._unk = value
-
-    @pad.setter
-    def pad(self, value):
-        self._pad = value
-
-    @eos.setter
-    def eos(self, value):
-        self._eos = value
-
-    @bos.setter
-    def bos(self, value):
-        self._bos = value
-
-    @property
-    def unk_id(self):
-        if self._unk is None:
-            return None
-        return self.convert_symbols_to_ids(self.eos)
-
-    @property
-    def pad_id(self):
-        if self._pad is None:
-            return None
-        return self.convert_symbols_to_ids(self.pad)
-
-    @property
-    def eos_id(self):
-        if self._eos is None:
-            return None
-        return self.convert_symbols_to_ids(self.eos)
-
-    @property
-    def bos_id(self):
-        if self._bos is None:
-            return None
-        return self.convert_symbols_to_ids(self.bos)
-    
     @abc.abstractmethod
     def get_one_sample(self, item):
         """Get one sample from dataset items.
-
         Args:
             item: one item in Dataset items.
                 Dataset items may include (raw_text, speaker_id, wav_path, ...)
-        
+
         Returns:
             sample (dict): sample dictionary return all feature used for preprocessing later.
         """
@@ -190,10 +128,10 @@ class BaseProcessor(metaclass=abc.ABCMeta):
         return sample
 
     @abc.abstractmethod
-    def text_to_sequence(self, text):
+    def text_to_sequence(self, text: str):
         return []
 
-    def convert_symbols_to_ids(self, symbols):
+    def convert_symbols_to_ids(self, symbols: Union[str, list]):
         sequence = []
         if isinstance(symbols, str):
             sequence.append(self._symbol_to_id[symbols])
@@ -203,10 +141,33 @@ class BaseProcessor(metaclass=abc.ABCMeta):
                 if isinstance(s, str):
                     sequence.append(self._symbol_to_id[s])
                 else:
-                    raise ValueError("All elements of symbols must be a str.")
+                    raise ValueError("All elements of symbols must be a string.")
         else:
             raise ValueError("A symbols must be a string or list of string.")
-        
+
         return sequence
 
+    def _load_mapper(self, loaded_path: str = None):
+        """
+        Save all needed mappers to file
+        """
+        loaded_path = os.path.join(self.data_dir, "mapper.json") if loaded_path is None else loaded_path
+        with open(loaded_path, "r") as f:
+            data = json.load(f)
+        self.speakers_map = data["speakers_map"]
+        self.symbol_to_id = data["symbol_to_id"]
+        self.id_to_symbol = {int(k): v for k, v in data["id_to_symbol"].items()}
 
+    def _save_mapper(self, saved_path: str = None):
+        """
+        Save all needed mappers to file
+        """
+        saved_path = os.path.join(self.data_dir, "mapper.json") if saved_path is None else saved_path
+        with open(saved_path, "w") as f:
+            full_mapper = {
+                "symbol_to_id": self.symbol_to_id,
+                "id_to_symbol": self.id_to_symbol,
+                "speakers_map": self.speakers_map,
+            }
+            json.dump(full_mapper, f)
+
