+++ tensorflow_tts/processor/kss.py
@@ -20,34 +20,43 @@ through Unidecode. For other data, you can modify _characters. See TRAINING_DATA
 from jamo import h2j, j2h
 from jamo.jamo import _jamo_char_to_hcj
 
-from tensorflow_tts.utils.korean import symbols, _symbol_to_id, _id_to_symbol
+from tensorflow_tts.utils.korean import symbols as KSS_SYMBOLS
+
+from dataclasses import dataclass
+from tensorflow_tts.processor import BaseProcessor
 
 # Regular expression matching text enclosed in curly braces:
 _curly_re = re.compile(r"(.*?)\{(.+?)\}(.*)")
 
-
-class KSSProcessor(object):
+@dataclass
+class KSSProcessor(BaseProcessor):
     """KSS processor."""
 
-    def __init__(self, data_dir, cleaner_names, metadata_filename="transcript.v.1.2.txt"):
-        self.data_dir = data_dir
-        self.cleaner_names = cleaner_names
+    cleaner_names: str = "korean_cleaners"
+    positions = {
+        "wave_file": 0,
+        "text_norm": 2,
+    }
+    train_f_name: str = "transcript.v.1.4.txt"
 
-        self.speaker_name = "kss"
-        if data_dir:
-            with open(os.path.join(data_dir, metadata_filename), encoding="utf-8") as f:
-                self.items = [self.split_line(data_dir, line, "|") for line in f]
+    def create_items(self):
+        if self.data_dir:
+            with open(os.path.join(self.data_dir, self.train_f_name), encoding="utf-8") as f:
+                self.items = [self.split_line(self.data_dir, line, "|") for line in f]
 
     def split_line(self, data_dir, line, split):
-        wav_file, _, text_norm, _, _ = line.strip().split(split)
-        wav_path = os.path.join(data_dir, wav_file)
-        return text_norm, wav_path, self.speaker_name
+        parts = line.strip().split(split)
+        wave_file = parts[self.positions["wave_file"]]
+        text_norm = parts[self.positions["text_norm"]]
+        wav_path = os.path.join(data_dir, "kss", wave_file)
+        speaker_name = "kss"
+        return text_norm, wav_path, speaker_name
     
     def get_one_sample(self, item):
-        text, wav_file, speaker_name = item
+        text, wav_path, speaker_name = item
 
         # normalize audio signal to be [-1, 1], soundfile already norm.
-        audio, rate = sf.read(wav_file)
+        audio, rate = sf.read(wav_path)
         audio = audio.astype(np.float32)
 
         # convert text to ids
@@ -57,7 +66,7 @@ class KSSProcessor(object):
             "raw_text": text,
             "text_ids": text_ids,
             "audio": audio,
-            "utt_id": os.path.split(wav_file)[-1].split(".")[0],
+            "utt_id": os.path.split(wav_path)[-1].split(".")[0],
             "speaker_name": speaker_name,
             "rate": rate,
         }
@@ -65,41 +74,43 @@ class KSSProcessor(object):
         return sample
 
     def text_to_sequence(self, text):
-        global _symbol_to_id
 
         sequence = []
         # Check for curly braces and treat their contents as ARPAbet:
         while len(text):
             m = _curly_re.match(text)
             if not m:
-                sequence += _symbols_to_sequence(
-                    _clean_text(text, [self.cleaner_names])
+                sequence += self._symbols_to_sequence(
+                    self._clean_text(text, [self.cleaner_names])
                 )
                 break
-            sequence += _symbols_to_sequence(
-                _clean_text(m.group(1), [self.cleaner_names])
+            sequence += self._symbols_to_sequence(
+                self._clean_text(m.group(1), [self.cleaner_names])
             )
-            sequence += _arpabet_to_sequence(m.group(2))
+            sequence += self._arpabet_to_sequence(m.group(2))
             text = m.group(3)
+
+        # add eos tokens
+        sequence += [self.eos_id]
         return sequence
 
 
-def _clean_text(text, cleaner_names):
-    for name in cleaner_names:
-        cleaner = getattr(cleaners, name)
-        if not cleaner:
-            raise Exception("Unknown cleaner: %s" % name)
-        text = cleaner(text)
-    return text
+    def _clean_text(self, text, cleaner_names):
+        for name in cleaner_names:
+            cleaner = getattr(cleaners, name)
+            if not cleaner:
+                raise Exception("Unknown cleaner: %s" % name)
+            text = cleaner(text)
+        return text
 
 
-def _symbols_to_sequence(symbols):
-    return [_symbol_to_id[s] for s in symbols if _should_keep_symbol(s)]
+    def _symbols_to_sequence(self, symbols):
+        return [self.symbol_to_id[s] for s in symbols if self._should_keep_symbol(s)]
 
 
-def _arpabet_to_sequence(text):
-    return _symbols_to_sequence(["@" + s for s in text.split()])
+    def _arpabet_to_sequence(self, text):
+        return self._symbols_to_sequence(["@" + s for s in text.split()])
 
 
-def _should_keep_symbol(s):
-    return s in _symbol_to_id and s != "_" and s != "~"
+    def _should_keep_symbol(self, s):
+        return s in self.symbol_to_id and s != "_" and s != "~"
